requirements.txt
aiogram==3.13.1
firebase-admin==6.5.0
beautifulsoup4==4.12.3
requests==2.32.3
apscheduler==3.10.4
python-dotenv==1.0.1
openai==1.51.2
httpx
beautifulsoup4
firebase-admin
python-dotenv
fake-useragent